requirements.txt
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
django-rest-framework==0.1.0
gunicorn==20.0.4
python-dotenv==0.12.0
django-countries==6.0
requests==2.23.0